+++ NEWS.md
@@ -4,6 +4,7 @@
 * refactor: Replace `tsk("boston_housing")` with `tsk("california_housing")`.
 * BREAKING CHANGE: Remove ``$loglik()`` method from all learners.
 * fix: Ignore `future.globals.maxSize` when `future::plan("sequential")` is used.
+* feat: Add `$characteristics` field to `Task` to store additional information.
 
 # mlr3 0.21.1